some extra gen level and lhe level information
